input: fix panic with unknown flags in input_verifier
this commit is from Android15:
    Add flags to input_verifier
    These flags are currently not known by the verifier. As a result, this
    causes a panic when such events are trying to get converted to the rust
    object.
    Add the missing flags in this CL.

Type: Fix
Redmine ID: 542835
Associated modifications: N/A
Test: setprop log.tag D

+++ libs/input/InputVerifier.cpp
@@ -41,7 +41,9 @@ Result<void> InputVerifier::processMovement(int32_t deviceId, int32_t action, ui
     rust::Slice<const RustPointerProperties> properties{rpp.data(), rpp.size()};
     rust::String errorMessage =
             android::input::verifier::process_movement(*mVerifier, deviceId, action, properties,
-                                                       flags);
+                                                       //------rk modiry start------
+                                                       static_cast<uint32_t>(flags));
+                                                       //------rk modiry end------
     if (errorMessage.empty()) {
         return {};
     } else {
